cloud_function/requirements.txt
functions-framework>=3.0.0
google-cloud-bigquery
requests